requirements.txt
torch>=1.12.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
pyyaml>=6.0
tqdm>=4.62.0
jupyter>=1.0.0
ipykernel>=6.0.0 
python-dotenv>=1.0.0
wandb>=0.16.0